split the model by files
